+++ stream-access-control/access-control-core/src/test/java/com/backbase/stream/service/AccessGroupServiceTest.java
@@ -1,16 +1,75 @@
 package com.backbase.stream.service;
 
-import com.backbase.dbs.accesscontrol.api.service.v2.*;
-import com.backbase.dbs.accesscontrol.api.service.v2.model.*;
+import static com.backbase.dbs.accesscontrol.api.service.v2.model.BatchResponseItemExtended.StatusEnum.HTTP_STATUS_INTERNAL_SERVER_ERROR;
+import static com.backbase.dbs.accesscontrol.api.service.v2.model.BatchResponseItemExtended.StatusEnum.HTTP_STATUS_OK;
+import static com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationAction.ADD;
+import static com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationAction.REMOVE;
+import static com.backbase.stream.legalentity.model.LegalEntityStatus.ENABLED;
+import static com.backbase.stream.test.LambdaAssertions.assertEqualsTo;
+import static com.backbase.stream.test.WebClientTestUtils.buildWebResponseExceptionMono;
+import static java.util.Arrays.asList;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.inOrder;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+import com.backbase.dbs.accesscontrol.api.service.v2.DataGroupApi;
+import com.backbase.dbs.accesscontrol.api.service.v2.DataGroupsApi;
+import com.backbase.dbs.accesscontrol.api.service.v2.FunctionGroupApi;
+import com.backbase.dbs.accesscontrol.api.service.v2.FunctionGroupsApi;
+import com.backbase.dbs.accesscontrol.api.service.v2.ServiceAgreementApi;
+import com.backbase.dbs.accesscontrol.api.service.v2.ServiceAgreementQueryApi;
+import com.backbase.dbs.accesscontrol.api.service.v2.ServiceAgreementsApi;
+import com.backbase.dbs.accesscontrol.api.service.v2.UserQueryApi;
+import com.backbase.dbs.accesscontrol.api.service.v2.UsersApi;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.BatchResponseItemExtended;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.FunctionGroupItem;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.FunctionGroupItem.TypeEnum;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.PersistenceApprovalPermissions;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.PersistenceApprovalPermissionsGetResponseBody;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationAction;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationAssignUserPermissions;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationFunctionGroupDataGroup;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationIdentifier;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationParticipantBatchUpdate;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationParticipantPutBody;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationServiceAgreementUserPair;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationServiceAgreementUsersBatchUpdate;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.ServiceAgreementItem;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.ServiceAgreementParticipantsGetResponseBody;
+import com.backbase.dbs.accesscontrol.api.service.v2.model.ServiceAgreementUsersQuery;
 import com.backbase.dbs.user.api.service.v2.UserManagementApi;
-import com.backbase.stream.legalentity.model.*;
+import com.backbase.stream.legalentity.model.BaseProductGroup;
+import com.backbase.stream.legalentity.model.BatchProductGroup;
+import com.backbase.stream.legalentity.model.BusinessFunctionGroup;
+import com.backbase.stream.legalentity.model.JobProfileUser;
+import com.backbase.stream.legalentity.model.LegalEntityParticipant;
+import com.backbase.stream.legalentity.model.ServiceAgreement;
+import com.backbase.stream.legalentity.model.ServiceAgreementUserAction;
+import com.backbase.stream.legalentity.model.User;
 import com.backbase.stream.product.task.BatchProductGroupTask;
 import com.backbase.stream.worker.exception.StreamTaskException;
 import com.backbase.stream.worker.model.StreamTask;
+import java.time.LocalDate;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import lombok.AllArgsConstructor;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
+import org.mockito.Answers;
+import org.mockito.ArgumentCaptor;
 import org.mockito.InOrder;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
@@ -20,24 +79,6 @@ import org.springframework.http.HttpMethod;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-
-import java.time.LocalDate;
-import java.util.*;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import static com.backbase.dbs.accesscontrol.api.service.v2.model.BatchResponseItemExtended.StatusEnum.HTTP_STATUS_INTERNAL_SERVER_ERROR;
-import static com.backbase.dbs.accesscontrol.api.service.v2.model.BatchResponseItemExtended.StatusEnum.HTTP_STATUS_OK;
-import static com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationAction.ADD;
-import static com.backbase.dbs.accesscontrol.api.service.v2.model.PresentationAction.REMOVE;
-import static com.backbase.stream.legalentity.model.LegalEntityStatus.ENABLED;
-import static com.backbase.stream.test.LambdaAssertions.assertEqualsTo;
-import static com.backbase.stream.test.WebClientTestUtils.buildWebResponseExceptionMono;
-import static java.util.Arrays.asList;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 class AccessGroupServiceTest {
@@ -74,6 +115,9 @@ class AccessGroupServiceTest {
 
     @Mock
     private ServiceAgreementsApi serviceAgreementsApi;
+
+    @Mock(answer = Answers.RETURNS_DEEP_STUBS)
+    private BackbaseStreamConfigurationProperties configurationProperties;
 
     @Test
     void getServiceAgreementByExternalIdRetrievesServiceAgreementByExternalId() {
@@ -590,6 +634,48 @@ class AccessGroupServiceTest {
         Assertions.assertSame(batchProductGroupTask, result);
 
         verify(accessControlUsersApi).putAssignUserPermissions(expectedPermissions);
+    }
+
+    @Test
+    void deleteFunctionGroupsForServiceAgreement_noneTypeConfigured_doesNotInvokeDeletion() {
+        String internalSaId = "sa-internal-id";
+
+        when(configurationProperties.getDbs().getDeletion().getFunctionGroupItemType()).thenReturn(FunctionGroupItemType.NONE);
+
+        subject.deleteFunctionGroupsForServiceAgreement(internalSaId).block();
+
+        verify(functionGroupsApi, times(0)).postFunctionGroupsDelete(any());
+    }
+
+    @Test
+    void deleteFunctionGroupsForServiceAgreement_templateTypeConfigured_deletesOnlyTemplateType() {
+        String internalSaId = "sa-internal-id";
+
+        FunctionGroupItem systemFunctionGroup = new FunctionGroupItem().id("system-group-id-1")
+            .name("SYSTEM_FUNCTION_GROUP")
+            .type(TypeEnum.SYSTEM);
+
+        FunctionGroupItem templateFunctionGroup = new FunctionGroupItem().id("template-group-id-2").name("Full access")
+            .type(TypeEnum.TEMPLATE);
+
+        when(functionGroupApi.getFunctionGroups(internalSaId))
+            .thenReturn(Flux.just(
+                systemFunctionGroup,
+                templateFunctionGroup
+            ));
+
+        when(functionGroupsApi.postFunctionGroupsDelete(any())).thenReturn(Flux.empty());
+
+        when(configurationProperties.getDbs().getDeletion().getFunctionGroupItemType()).thenReturn(FunctionGroupItemType.TEMPLATE);
+
+        subject.deleteFunctionGroupsForServiceAgreement(internalSaId).block();
+
+        ArgumentCaptor<List<PresentationIdentifier>> captor = ArgumentCaptor.forClass(
+            List.class);
+        verify(functionGroupsApi).postFunctionGroupsDelete(captor.capture());
+
+        List<PresentationIdentifier> value = captor.getValue();
+        assertEquals(templateFunctionGroup.getId(), value.get(0).getIdIdentifier());
     }
 
     private void thenRegularUsersUpdateCall(String expectedSaExId, PresentationAction expectedAction,